fix: Backport CAMEL-16556 to the Camel2 camel-atlasmap component
Fixes: #2708
Moved propertiesFile parameter from endpoint to component
